stop collecting plans after deadline has been exceeded
Set a collection deadline to make sure we stop collecting plans when it's time for the next collection.

If we find a large number of plans during a single collection run and the loading of those plans takes longer than usual, the total time required to collect all of the plans can take longer than the collection interval, which means the following run of the interval will be delayed.

This ensures we stop collecting plans in time for the next collection run.